requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
twikit>=0.3.0
feedgen>=0.9.0
python-dotenv>=1.0.0
requests>=2.31.0
